requirements.txt
streamlit
plotly
pandas
pyarrow
python-dotenv
numpy
